# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SSC GD Smart Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/ask:
    post:
      operationId: askAI
      tags: [ai]
      summary: Ask AI a doubt or question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIAskInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
  /ai/explain:
    post:
      operationId: explainQuestion
      tags: [ai]
      summary: Get AI explanation for a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIExplainInput"
      responses:
        "200":
          description: AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
  /ai/study-plan:
    post:
      operationId: generateStudyPlan
      tags: [ai]
      summary: Generate personalized study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /ai/analyze-test:
    post:
      operationId: analyzeTestResult
      tags: [ai]
      summary: AI analysis of test performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestAnalysisInput"
      responses:
        "200":
          description: AI analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
  /ai/chat:
    post:
      operationId: chatWithAI
      tags: [ai]
      summary: Chat with AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatInput"
      responses:
        "200":
          description: AI chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: isPremium
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/tests:
    post:
      operationId: createTest
      tags: [admin]
      summary: Create a new test (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestInput"
      responses:
        "201":
          description: Created test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Test"
  /admin/questions:
    post:
      operationId: createQuestion
      tags: [admin]
      summary: Create a new question (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /admin/notes:
    post:
      operationId: createNote
      tags: [admin]
      summary: Create/upload a note (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /admin/videos:
    post:
      operationId: createVideo
      tags: [admin]
      summary: Create a video entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /admin/notifications:
    post:
      operationId: sendNotification
      tags: [admin]
      summary: Send notification to users (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, email, role, isPremium, joinedAt]
      properties:
        id:          { type: string }
        name:        { type: string }
        email:       { type: string }
        phone:       { type: ["string", "null"] }
        avatar:      { type: ["string", "null"] }
        role:        { type: string, enum: [student, admin] }
        isPremium:   { type: boolean }
        premiumExpiresAt: { type: ["string", "null"] }
        totalPoints: { type: integer }
        rank:        { type: integer }
        joinedAt:    { type: string }
        studyHours:  { type: number }
        targetExam:  { type: string }
...
    Test:
      type: object
      required: [id, title, type, subject, totalQuestions, duration, maxMarks, negativeMarking, isPremium, attempts]
      properties:
        id:              { type: string }
        title:           { type: string }
        type:            { type: string, enum: [full, chapter, pyq] }
        subject:         { type: ["string", "null"] }
        totalQuestions:  { type: integer }
        duration:        { type: integer }
        maxMarks:        { type: number }
        negativeMarking: { type: number }
        isPremium:       { type: boolean }
        attempts:        { type: integer }
        year:            { type: ["integer", "null"] }
        description:     { type: ["string", "null"] }
        createdAt:       { type: string }
...
    TestInput:
      type: object
      required: [title, type, duration, negativeMarking]
      properties:
        title:           { type: string }
        type:            { type: string, enum: [full, chapter, pyq] }
        subject:         { type: string }
        duration:        { type: integer }
        negativeMarking: { type: number }
        isPremium:       { type: boolean }
        year:            { type: integer }
        description:     { type: string }

    # ─── Question ───────────────────────────────────────────────────────────
    Question:
      type: object
      required: [id, text, options, correctOption, subject, difficulty]
      properties:
        id:             { type: string }
        text:           { type: string }
        options:
          type: array
          items:
            type: object
            required: [id, text]
            properties:
              id:   { type: string }
              text: { type: string }
        correctOption:  { type: string }
        explanation:    { type: ["string", "null"] }
        subject:        { type: string }
        topic:          { type: ["string", "null"] }
        difficulty:     { type: string, enum: [easy, medium, hard] }
        year:           { type: ["integer", "null"] }
        isPYQ:          { type: boolean }
        isBookmarked:   { type: boolean }
...
    QuestionInput:
      type: object
      required: [text, options, correctOption, subject, difficulty]
      properties:
        text:
          type: string
        options:
          type: array
          items:
            type: object
            required: [id, text]
            properties:
              id:   { type: string }
              text: { type: string }
        correctOption: { type: string }
        explanation:   { type: string }
        subject:       { type: string }
        topic:         { type: string }
        difficulty:    { type: string, enum: [easy, medium, hard] }
        year:          { type: integer }
        isPYQ:         { type: boolean }
    BookmarkStatus:
      type: object
      required: [success, isBookmarked]
      properties:
        success:    { type: boolean }
        isBookmarked: { type: boolean }

    # ─── Note ───────────────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, title, subject, chapter, fileUrl, isPremium, downloadCount]
      properties:
        id:            { type: string }
        title:         { type: string }
        subject:       { type: string }
        chapter:       { type: string }
        description:   { type: ["string", "null"] }
        fileUrl:       { type: string }
        thumbnailUrl:  { type: ["string", "null"] }
        fileSize:      { type: ["string", "null"] }
        isPremium:     { type: boolean }
        downloadCount: { type: integer }
        isBookmarked:  { type: boolean }
        createdAt:     { type: string }
    NoteInput:
      type: object
      required: [title, subject, chapter, fileUrl]
      properties:
        title:        { type: string }
        subject:      { type: string }
        chapter:      { type: string }
        description:  { type: string }
        fileUrl:      { type: string }
        thumbnailUrl: { type: string }
        fileSize:     { type: string }
        isPremium:    { type: boolean }
...
    Video:
      type: object
      required: [id, title, subject, duration, isPremium, thumbnailUrl, views]
      properties:
        id:           { type: string }
        title:        { type: string }
        subject:      { type: string }
        topic:        { type: ["string", "null"] }
        description:  { type: ["string", "null"] }
        videoUrl:     { type: ["string", "null"] }
        thumbnailUrl: { type: string }
        duration:     { type: integer }
        isPremium:    { type: boolean }
        instructor:   { type: string }
        views:        { type: integer }
        progress:     { type: ["number", "null"] }
        createdAt:    { type: string }
...
    VideoInput:
      type: object
      required: [title, subject, videoUrl, thumbnailUrl, duration, instructor]
      properties:
        title:        { type: string }
        subject:      { type: string }
        topic:        { type: string }
        description:  { type: string }
        videoUrl:     { type: string }
        thumbnailUrl: { type: string }
        duration:     { type: integer }
        isPremium:    { type: boolean }
        instructor:   { type: string }

    # ─── AI ─────────────────────────────────────────────────────────────────
    AIAskInput:
      type: object
      required: [question]
      properties:
        question: { type: string }
        subject:  { type: string }
        context:  { type: string }
    AIExplainInput:
      type: object
      required: [questionId]
      properties:
        questionId: { type: string }
        language:   { type: string, enum: [english, hindi] }
    StudyPlanInput:
      type: object
      required: [targetDate, weakSubjects, studyHoursPerDay]
      properties:
        targetDate:        { type: string }
        weakSubjects:
          type: array
          items:
            type: string
        studyHoursPerDay:  { type: number }
        currentLevel:      { type: string, enum: [beginner, intermediate, advanced] }
    TestAnalysisInput:
      type: object
      required: [testResultId]
      properties:
        testResultId: { type: string }
    AIChatInput:
      type: object
      required: [message]
      properties:
        message:        { type: string }
        conversationId: { type: ["string", "null"] }
    AIResponse:
      type: object
      required: [response, conversationId]
      properties:
        response:       { type: string }
        conversationId: { type: string }
        suggestions:
          type: array
          items:
            type: string
...
    StudyPlan:
      type: object
      required: [plan, totalWeeks, dailyHours]
      properties:
        plan:       { type: string }
        totalWeeks: { type: integer }
        dailyHours: { type: number }
        weeks:
          type: array
          items:
            type: object
            required: [week, topics, goals]
            properties:
              week:   { type: integer }
              topics:
                type: array
                items:
                  type: string
              goals:
                type: array
                items:
                  type: string

    # ─── Notifications ───────────────────────────────────────────────────────
...
    NotificationInput:
      type: object
      required: [title, message, type]
      properties:
        title:   { type: string }
        message: { type: string }
        type:    { type: string, enum: [exam, reminder, result, general] }
        target:  { type: string, enum: [all, premium, free] }

    # ─── Admin ───────────────────────────────────────────────────────────────
...
    PaginatedUsers:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }